ajout des stats
