+++ src/app/features/belote/game/round/drawer/last-turn/last-turn.component.scss
@@ -2,11 +2,6 @@
   display: block;
 }
 
-h3 {
-  text-align: center;
-  font-weight: bold;
-}
-
 .cards {
   display: flex;
   width: 100%;
