+++ src/app/features/belote/game/round/drawer/stats/stats.component.scss
@@ -1,0 +1,27 @@
+:host {
+  display: block;
+  max-width: 500px;
+  width: 80%;
+  margin: 0 auto;
+}
+
+table {
+  width: 100%;
+  text-align: center;
+  table-layout: fixed;
+  border-collapse: collapse;
+}
+
+tr:nth-child(2n + 1) td {
+  background-color: rgba(200, 200, 200, 0.3);
+}
+
+tr:first-child th {
+  border-bottom: 1px solid black;
+}
+
+th:first-child,
+td:first-child {
+  border-right: 1px solid black;
+  width: 50%;
+}
