+++ src/app/features/belote/game/round/drawer/stats/stats.module.ts
@@ -1,0 +1,10 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { StatsComponent } from './stats.component';
+
+@NgModule({
+  declarations: [StatsComponent],
+  imports: [CommonModule],
+  exports: [StatsComponent],
+})
+export class StatsModule {}
