+++ src/styles/_typo.scss
@@ -1,3 +1,9 @@
+
+h3 {
+  text-align: center;
+  font-weight: bold;
+}
+
 p {
   margin: 0;
   text-align: center;
